requirements.txt
numpy==1.19.2  # with pip: in general, use pip where possible
torch==1.8.0
torchvision==0.9.0
piqa==1.1.0
pyyaml==5.4.1
yaml~=0.2.5
scipy==1.5.2 # try conda if weird error happens
matplotlib==3.4.1
joblib==0.17.0
pandas==1.1.3
scikit-learn==0.23.2
scikit-image==0.17.2
seaborn==0.11.1
statsmodels==0.12.2
lime==0.2.0.1
shap==0.39.0
pillow==8.0.1
tensorboard==2.3.0
